+++ packages/video_player/video_player/CHANGELOG.md
@@ -1,3 +1,7 @@
+## NEXT
+
+* Updated Android lint settings.
+
 ## 2.1.14
 
 * Removed dependency on the `flutter_test` package.